memory calculated per model

+++ classify.py
@@ -287,6 +287,7 @@ def main(size='sm', file2='0', file3='0', red='0', tfidf='0', prune='0', est='32
         tfer = TfidfTransformer()
         tfer.fit(features)
         features = tfer.transform(features)
+        features = features.astype('float16')
 
     print "Final data shape:", features.shape
     logging.info("Final data shape: %s" % (features.shape,))
@@ -304,7 +305,7 @@ def main(size='sm', file2='0', file3='0', red='0', tfidf='0', prune='0', est='32
         print "Time elapsed for dimensionality reduction is %f" %  elapsed
         logging.info("Time elapsed for dimensionality reduction is %f" %  elapsed)
 
-    #features = features.astype('float16')
+
 
     results = classify_all(labels, features, clfs, folds, model_names)
     #results.sort("Split Val Acc", inplace=True, ascending=False)
